requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-snowflake==5.1.0
pyspark==3.5.0
delta-spark==3.0.0
dbt-core==1.7.0
dbt-snowflake==1.7.0
great-expectations==0.18.0
snowflake-connector-python==3.6.0
pandas==2.1.4
pyarrow==14.0.2
boto3==1.34.0
pyyaml==6.0.1
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
